Renamed InvalidDataPath to InvalidDataArchive to match other archives. Moved archive related classes into their own files
