+++ AnnoMapEditor/DataArchives/DataArchive.cs
@@ -1,13 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
+﻿using System.IO;
 using System.Threading.Tasks;
 
 namespace AnnoMapEditor.DataArchives
 {
     public static class DataArchive
     {
-        public static readonly IDataArchive Default = new InvalidDataPath("");
+        public static readonly IDataArchive Default = new InvalidDataArchive("");
 
         public static async Task<IDataArchive> OpenAsync(string? folderPath)
         {
@@ -42,38 +40,6 @@ namespace AnnoMapEditor.DataArchives
             if (Directory.Exists(Path.Combine(path, "dlc01")))
                 return Path.GetDirectoryName(path);
             return null;
-        }
-    }
-
-    public interface IDataArchive
-    {
-        bool IsValid { get; }
-        string Path { get; }
-
-        Stream? OpenRead(string filePath);
-        IEnumerable<string> Find(string pattern);
-        Task LoadAsync();
-    }
-
-    public class InvalidDataPath : IDataArchive
-    {
-        public bool IsValid { get; } = false;
-        public Stream? OpenRead(string filePath) => null;
-        public string Path { get; }
-
-        public InvalidDataPath(string path)
-        {
-            Path = path;
-        }
-
-        public Task LoadAsync()
-        {
-            return Task.Run(() => { });
-        }
-
-        public IEnumerable<string> Find(string pattern)
-        {
-            return Array.Empty<string>();
         }
     }
 }
